feat: update memory addresses
- Using Long instead of Integer;
- Memory addresses are usually reffered as unsigned integers;

+++ src/main/java/br/ufrpe/mips/data/IMemoryLocation.java
@@ -29,7 +29,7 @@ public interface IMemoryLocation<T> {
    * 
    * @return inteiro, indicando o endereço dessa localização de memória.
    */
-  int address();
+  long address();
 
   /**
    * Qual valor armazenado nessa localização de memória.
